Add file analysis
This commit adds the functionality to analyse the files from the glob.
Each file is analysed in its own instance of the CodeAnalyzer and reports
back to the JSCC module. This module keeps a count of the number of
outstanding calls, once it reaches zero we know the work is done.