Check files and generate basic report
The entire check for API has been changed to better suit the Node.js
flow.

Instead of loading the compatibility data manually this is now done
by simply requiring the two JSON files. This should work on Node v0.5
and later.

Reading the files from the system is now handled by vow-fs. This is a
promise based way of reading the files.

The console reporter simply outputs the detected APIs and the files in
which the APIs were detected.